+++ metrics/middleware.py
@@ -22,12 +22,16 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
 # THE POSSIBILITY OF SUCH DAMAGE.
 
+import logging
+
 from django.core.exceptions import ValidationError
 from django.utils.deprecation import MiddlewareMixin
 
 from . import settings
 from .models import Request
 from .router import Patterns
+
+logger = logging.getLogger("metrics.security.middleware")
 
 
 class RequestMiddleware(MiddlewareMixin):
@@ -57,8 +61,13 @@ class RequestMiddleware(MiddlewareMixin):
         try:
             r.from_http_request(request, response, commit=False)
             r.full_clean()
-        except ValidationError:
-            pass
+        except ValidationError as exc:
+            logger.warning(
+                "Bad request: %s",
+                str(exc),
+                exc_info=exc,
+                extra={"status_code": 400, "request": request},
+            )
         else:
             r.save()
         return response